cloudbuild-and-automation/scripts/requirements.txt
# Function dependencies, for example:
# package>=version
google-auth
google-cloud-containeranalysis
google-cloud-build
envs
